GH-79 repackage content parser and further tiddy up.
