+++ logging.properties
@@ -28,7 +28,7 @@ java.util.logging.ConsoleHandler.level=INFO
 #java.util.logging.ConsoleHandler.level = INFO
 # Set the default logging level for new FileHandler instances
 java.util.logging.FileHandler.level=OFF
-org.icepdf.core.util.content.level=INFO
+org.icepdf.core.util.parser.content.level=INFO
 org.icepdf.core.pobjects.fonts.level=INFO
 org.icepdf.core.pobjects.fonts.nfont.Font.level=INFO
 org.icepdf.level=FINER
